Added three notebooks with several dSA-related studies. These should serve as a baseline for other muon efficiency studies too, in particular displaced (e.g. displaced global muons).
Also added several configs to go along with the notebooks, and some to run via the tnp_fitter script (those
are labeled with a '_official' suffix). There are both ID and reconstruction configs.

Finally, there are several new registries to locate the required ntuples, which were produced with the new
ntuplizer by the Muon POG.